+++ components/asic/test/test_timeout.c
@@ -5,10 +5,10 @@
 TEST_CASE("Check asic timeout 1x BM1397", "[common]")
 {
     float frequency = 450.0; // MHz
-    uint16_t asic_count = 1;
-    uint16_t small_cores = 672;
-    uint16_t cores = 168;
-    float version_size = 4.0;
+    size_t asic_count = 1;
+    size_t small_cores = 672;
+    size_t cores = 168;
+    size_t version_size = 4;
     float timeout_percent = 0.75;
     float default_timeout_ms = 20;
 
@@ -21,10 +21,10 @@ TEST_CASE("Check asic timeout 1x BM1397", "[common]")
 TEST_CASE("Check asic timeout 2x BM1370", "[common]")
 {
     float frequency = 450.0; // MHz
-    uint16_t asic_count = 2;
-    uint16_t small_cores = 2040;
-    uint16_t cores = 128;
-    float version_size = 65536.0;
+    size_t asic_count = 2;
+    size_t small_cores = 2040;
+    size_t cores = 128;
+    size_t version_size = 65536;
     float timeout_percent = 0.5;
     float default_timeout_ms = 500;
 
@@ -37,10 +37,10 @@ TEST_CASE("Check asic timeout 2x BM1370", "[common]")
 TEST_CASE("Check default asic timeout 0x BM1370", "[common]")
 {
     float frequency = 450.0; // MHz
-    uint16_t asic_count = 0; // 0 chip example
-    uint16_t small_cores = 2040;
-    uint16_t cores = 128;
-    float version_size = 65536.0;
+    size_t asic_count = 0; // 0 chip example
+    size_t small_cores = 2040;
+    size_t cores = 128;
+    size_t version_size = 65536;
     float timeout_percent = 0.5;
     float default_timeout_ms = 500;
 
@@ -52,10 +52,10 @@ TEST_CASE("Check default asic timeout 0x BM1370", "[common]")
 TEST_CASE("Check asic timeout 3x BM1370", "[common]")
 {
     float frequency = 450.0; // MHz
-    uint16_t asic_count = 3; // not power of 2 chain length
-    uint16_t small_cores = 2040;
-    uint16_t cores = 128;
-    float version_size = 256.0;
+    size_t asic_count = 3; // not power of 2 chain length
+    size_t small_cores = 2040;
+    size_t cores = 128;
+    size_t version_size = 256;
     float timeout_percent = 0.5;
     float default_timeout_ms = 500;
 
@@ -68,10 +68,10 @@ TEST_CASE("Check asic timeout 3x BM1370", "[common]")
 TEST_CASE("Check max asic timeout 1x BM1370", "[common]")
 {
     float frequency = 450.0; // MHz
-    uint16_t asic_count = 1;
-    uint16_t small_cores = 2040;
-    uint16_t cores = 128;
-    float version_size = 65536.0;
+    size_t asic_count = 1;
+    size_t small_cores = 2040;
+    size_t cores = 128;
+    size_t version_size = 65536;
     float timeout_percent = 1.0;
     float default_timeout_ms = 500;
